requirements.txt
# Updated requirements
pandas
librosa
pydub

SpeechRecognition
torch

transformers
google-generativeai

pronouncing

IPython
